refactor: chart folder structure
