+++ src/components/common/NE_Chart/ChartApexArea/index.js
@@ -1,0 +1,1 @@
+export { default } from './apex_area';
