+++ src/components/common/NE_Chart/ChartApexPie/index.js
@@ -1,0 +1,1 @@
+export { default } from './apex_pie';
